backend/user_ms/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
PyJWT==2.8.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
bcrypt==4.1.2
flask-cors==4.0.0
pika==1.3.2
